% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSIT_est.R
\name{RSIT_est}
\alias{RSIT_est}
\title{NULL model estimation}
\usage{
RSIT_est(Y, X, E, family = "Guassian", GRM = NULL)
}
\arguments{
\item{Y}{phenotype vector}

\item{X}{Covariate matrix denoting the covariate variables}

\item{E}{environment vector}

\item{family}{String, the phenotype model, two optional values: 'Guassian', 'binomial'}

\item{GRM}{takes m-by-m genetic correlation matrix or kinship matrix.}
}
\value{
This function returns a list object with model parameters and residuals of the NULL model
}
\description{
This function estimate the parameters and residuals for the NULL model in RSIT
}
