% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSIT_simu.R
\name{simu.binary.phe}
\alias{simu.binary.phe}
\title{Simulation for RSIT test}
\usage{
simu.binary.phe(
  n.sample,
  geno,
  n_casual = c(2, 8),
  main_effect = c(0.1, 0.2),
  GE_effect = c(0.3, 0.88),
  direction = c("same", "same"),
  GEindependent = T,
  power = T,
  maineffect = T,
  SNPtype = "rare_common",
  main_GE = T
)
}
\arguments{
\item{n.sample}{Numeric, sample size, number of individuals}

\item{geno}{matrix, genotype matrix}

\item{n_casual}{Vector, number of causal SNPs for common SNPs and rare SNPs}

\item{main_effect}{Vector, gene-environment interaction effect for common SNPs and rare SNPs}

\item{direction}{Vector, main effect and gene-environment interaction effect direction}

\item{GEindependent}{Logical variable, indicating whether gene and environment variable are independent}

\item{power}{Logical variable, indicating whether include disease genes in the generated genes}

\item{maineffect}{Logical variable, indicating whether include main effect}
}
\value{
A list object is returned to be used as object for RSIT test
}
\description{
This function use pre-defined parameters to make the simulation data for the RSIT test (including type I and power test)
}
