% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSIT_simu.R
\name{simu_snp_mat}
\alias{simu_snp_mat}
\title{Simulate the genotype data}
\usage{
simu_snp_mat(n.sample, mat.count, common_snp = F, independent = F)
}
\arguments{
\item{n.sample}{Numeric, sample size, number of individuals}

\item{mat.count}{Numeric, number of SNP regions}

\item{common_snp}{Logical variable, indicating whether  the region include common SNPs}

\item{independent}{Logical variable, indicating whether SNPs in the same region are independent}
}
\value{
A list object is returned to be used as genotype for RSIT test
}
\description{
This function use pre-defined parameters to simulate the genotype data
}
